stop executor from scheduled task; retry v2,v3 metrics table get logic to handle exception

+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/process/MetricsTableDeleter.java
@@ -16,6 +16,7 @@
 package co.cask.cdap.metrics.process;
 
 import co.cask.cdap.common.conf.CConfiguration;
+import co.cask.cdap.data2.dataset2.lib.table.MetricsTable;
 import co.cask.cdap.data2.util.hbase.HBaseDDLExecutorFactory;
 import co.cask.cdap.data2.util.hbase.HBaseTableUtil;
 import co.cask.cdap.metrics.store.MetricDatasetFactory;
@@ -26,32 +27,61 @@ import org.slf4j.LoggerFactory;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 
 /**
  * Deletes old v2 metrics tables when they no longer have any data. ie - data migration to new v3 table is complete
  */
-public class MetricsTableDeleter implements Runnable {
+public class MetricsTableDeleter {
   private static final Logger LOG = LoggerFactory.getLogger(MetricsTableDeleter.class);
 
   private final Map<Integer, MetricsTableMigration> resolutionTablesToDelete;
   private final HBaseTableUtil hBaseTableUtil;
   private final HBaseDDLExecutorFactory hBaseDDLExecutorFactory;
+  private final ScheduledExecutorService scheduledExecutorService;
 
   public MetricsTableDeleter(MetricDatasetFactory metricDatasetFactory, CConfiguration cConf, Configuration hConf,
-                             HBaseTableUtil hBaseTableUtil, List<Integer> resolutions) {
+                             HBaseTableUtil hBaseTableUtil, List<Integer> resolutions,
+                             ScheduledExecutorService scheduledExecutorService) {
     resolutionTablesToDelete = new HashMap<>();
     for (int resolution : resolutions) {
-      resolutionTablesToDelete.put(resolution,
-                                   new MetricsTableMigration(metricDatasetFactory, resolution, cConf, hConf));
+      MigrationTableUtility migrationTableUtility = new MigrationTableUtility(cConf, hConf,
+                                                                              metricDatasetFactory, hBaseTableUtil,
+                                                                              resolution);
+      if (migrationTableUtility.v2MetricsTableExists()) {
+        MetricsTable v3MetricsTable = migrationTableUtility.getV3MetricsTable();
+        MetricsTable v2MetricsTable = migrationTableUtility.getV2MetricsTable();
+        resolutionTablesToDelete.put(resolution,
+                                     new MetricsTableMigration(v2MetricsTable, v3MetricsTable, cConf, hConf));
+      }
     }
     this.hBaseTableUtil = hBaseTableUtil;
     this.hBaseDDLExecutorFactory = new HBaseDDLExecutorFactory(cConf, hConf);
+    this.scheduledExecutorService = scheduledExecutorService;
   }
+
+  /**
+   * if the metrics data migration  and all tables are deleted, we would just shutdown the executor service and return
+   * if not, we would schedule deletion task at fixed rate schedule.
+   * when the deletion of tables is complete, the task would shutdown the executor service by itself
+   */
+  public void scheduleDeletionIfNecessary() {
+    if (!allTablesDeleted()) {
+      scheduledExecutorService.scheduleAtFixedRate(new DeletionTask(), 1, 2, TimeUnit.HOURS);
+      LOG.info("Scheduled metrics deletion thread for 1, 60, 3600, INT_MAX resolution tables, " +
+                 "tables will only be deleted after data migration is completed for them");
+    } else {
+      LOG.debug("Not Scheduling Deletion thread, all tables have already been deleted");
+      scheduledExecutorService.shutdown();
+    }
+  }
+
   /**
    * returns true when data migration is complete and all v2 metrics tables have been deleted; false otherwise
    * @return true if migration is complete; false otherwise
    */
-  public boolean allTablesDeleted() {
+  private boolean allTablesDeleted() {
     // when metrics.processor restarts and initializes TableDeleter,
     // this is useful to decide whether to schedule deletion thread or not
     for (Map.Entry<Integer, MetricsTableMigration> entry : resolutionTablesToDelete.entrySet()) {
@@ -63,15 +93,27 @@ public class MetricsTableDeleter implements Runnable {
     return true;
   }
 
-  @Override
-  public void run() {
-    for (int resolution : resolutionTablesToDelete.keySet()) {
-      MetricsTableMigration metricsTableMigration = resolutionTablesToDelete.get(resolution);
-      if (metricsTableMigration.v2MetricsTableExists(hBaseTableUtil, resolution) &&
-        !metricsTableMigration.isOldMetricsDataAvailable() &&
-        metricsTableMigration.deleteV2MetricsTable(hBaseDDLExecutorFactory, hBaseTableUtil, resolution)) {
-        LOG.info("Successfully Deleted the v2 metrics table for resolution {} seconds", resolution);
+  private final class DeletionTask implements Runnable {
+    @Override
+    public void run() {
+
+      if (allTablesDeleted()) {
+        // this log will be printed only once when all tables are deleted
+        // and the future runs will be skipped earlier before scheduling the Deletion task
+        LOG.info("All Metrics tables have been deleted successfully");
+        scheduledExecutorService.shutdown();
+        return;
+      }
+
+      for (int resolution : resolutionTablesToDelete.keySet()) {
+        MetricsTableMigration metricsTableMigration = resolutionTablesToDelete.get(resolution);
+        if (metricsTableMigration.v2MetricsTableExists(hBaseTableUtil, resolution) &&
+          !metricsTableMigration.isOldMetricsDataAvailable() &&
+          metricsTableMigration.deleteV2MetricsTable(hBaseDDLExecutorFactory, hBaseTableUtil, resolution)) {
+          LOG.info("Successfully Deleted the v2 metrics table for resolution {} seconds", resolution);
+        }
       }
     }
   }
+
 }
